+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.4.0] - 2022-11-27
+
 ### Added
 
 - JSON result will have the plan run duration on `run_duration` field.
@@ -44,7 +46,8 @@
 - Different exit codes depending on the result of the run.
 - Dry-run mode.
 
-[unreleased]: https://github.com/slok/tfe-drift/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/slok/tfe-drift/compare/v0.4.0...HEAD
+[v0.4.0]: https://github.com/slok/tfe-drift/compare/v0.3.0...v0.4.0
 [v0.3.0]: https://github.com/slok/tfe-drift/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slok/tfe-drift/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/slok/tfe-drift/releases/tag/v0.1.0